♻️ inline e2e playwright base config into main config
- Delete test/e2e/playwright.base.config.ts; merge its contents into
  test/e2e/playwright.config.ts so there is a single source of truth.
- Drop the hardcoded --project=chromium from the test:e2e script so it
  accepts a --project flag from callers.
- Update .gitlab-ci.yml and docs/TESTING.md to invoke
  `yarn test:e2e --project=$BROWSER` instead of the long
  `yarn playwright test --config ...` form.
